+++ Seat.h
@@ -34,7 +34,7 @@ namespace Airport {
 
         virtual void writeToFile(ostream& output);
         virtual void printSeat();
-        virtual double calcPrice(double days);
+        virtual double calcPrice(int days);
 
     private:
         Flight* theFlight;
